Update README.md
The Multiple Probability Calculator is a versatile desktop application designed to calculate probabilities and analyze statistical distributions for various use cases. This tool is ideal for students, educators, researchers, and professionals, offering a comprehensive range of calculators, intuitive design, and educational value.

Important Note for Downloaders
As a solo developer, the application is not digitally signed, which may cause antivirus software to flag it as potentially malicious. To ensure your safety, scan the file with your antivirus software before unzipping or running it. This helps verify the file’s integrity and provides reassurance.

All Instructions to utilize the program are included within the readme file.